fixing more folder stuff. Prisma is initiated and schema moved to /app/lib/prisma
